% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Horseshoesp.R
\name{horseshoesp}
\alias{horseshoesp}
\title{Horseshoe Sparse Regression}
\usage{
horseshoesp(
  y,
  X,
  method.tau = c("fixed", "truncatedCauchy", "halfCauchy"),
  tau = 1,
  method.sigma = c("fixed", "Jeffreys"),
  Sigma2 = 1,
  s = 0.5,
  burn = 1000,
  nmc = 5000,
  thin = 1,
  alpha = 0.05
)
}
\arguments{
\item{y}{Response vector.}

\item{X}{Sparse design matrix}
}
\value{
A fitted model object.
}
\description{
This function performs Horseshoe regression optimized for sparse matrices.
}
